company name is Nitin international.
it looks okayish. most of the pages do not work.
was hoping web site to be attractive and have stock photos. the font and color schemes look okayish nothing pretty about iut.

the website is bland and looks more like a pdf document or ppt .

the home page should be attractive