About Us
We, Nitin International an overseas manpower consultant , Govt Approved recognized recruitment agency and have earned international reputation. We follow a well-established process to screen our candidates which is extensive and thorough. We meticulously provide background checks and references, both academic as well as professional. We value quality and trust as the foundation of our company.

With a team of dedicated people, we understand your requirements and find the best-possible candidates. We provide the best solutions to your problems. Most importantly, we can handle multiple projects and the team always deliver within the deadline. As professionals, we realize that time is valuable to employers looking for a trained and professional staff.

Vision & Mission
“We seek to enhance the life experiences of our candidates and clients by coordinating a valuable recruiting experience.” Our mission is to ensure expert recruitment across various services and help our clients secure the best talent.

In a well-coordinated effort, the entire recruitment process enhances the hiring experience and reduces wastage of time. We constantly strive to meet established expectations of clients and candidates, and maintain work ethics.

Even after recruitment, we keep in constant touch with the candidate to help him integrate into a new company and work culture. We serve to satisfy the people we work with.

Core Values
Our reputation and business is built on the values we practice. They form the foundations of our company. Approach: Our clients’ interest comes first. We give a competitive advantage to our clients by first understanding their needs, and then by delivering the best.

Integrity: For a company, the integrity of the candidates they will be hiring is of utmost importance. They trust us to screen and evaluate in order to match the profiles. We ensure this through a multi-step assessment process.

Respect: The core nature of our business is to liaison between candidates and clients. We respect both parties’ wishes and demands.
The Founder

Mr. Vimal Kumar Jaitly
The Founder
Mr. Vimal Kumar Jaitly who has over 35 years of experience in the travel & trade industries with leading organizations in India, this company has established itself as a reliable pool of resources within a short span of time. From finding a candidate, holding interviews to even providing the visas, Nitin International has been setting an example of exemplary work ethic since 1993.

Mr. Jaitly has successfully guided the company by building strong foundations of quality, punctuality, dedication and perfectionism. To him, there is no room for error, specially when it comes to job seekers’ and employers’ satisfaction.

Authorized by
“Royal Embassy of Saudi Arabia” New Delhi, India

“Embassy of the state of Kuwait” Delhi, India (for dealing with all visa related matters and clearing of visa)

Licence: We are registered with Ministry of Overseas Indian Affairs,Government of India for manpower recruitment, Registration No: B-0717/DEL/PER/1000+/3/3674/93 valid until 4th Aug, 2018